+++ core/moengage_flutter_android/lib/moengage_flutter_android.dart
@@ -325,4 +325,9 @@ class MoEngageFlutterAndroid extends MoEngageFlutterPlatform {
       return Future.error(exception);
     }
   }
+
+  @override
+  void registerForProvisionalPush() {
+    Logger.v('registerForProvisionalPush() method not available for Android');
+  }
 }