+++ core/moengage_flutter_ios/lib/moengage_flutter_ios.dart
@@ -252,4 +252,9 @@ class MoEngageFlutterIOS extends MoEngageFlutterPlatform {
       return Future.error(exception);
     }
   }
+
+  @override
+  void registerForProvisionalPush() {
+    _channel.invokeMethod(methodiOSRegisterProvisionalPush);
+  }
 }