+++ core/moengage_flutter/CHANGELOG.md
@@ -6,6 +6,8 @@
 - Added support for multiple self-handled in-apps
 - Android
     - Bundling `moe-android-sdk` & `inapp` dependencies with the plugin
+    - `moe-android-sdk` version updated to `13.05.00`
+    - `inapp` version updated to `8.6.0`
 
 # 07-08-2024