+++ core/moengage_flutter/CHANGELOG.md
@@ -8,6 +8,8 @@
     - Bundling `moe-android-sdk` & `inapp` dependencies with the plugin
     - `moe-android-sdk` version updated to `13.05.00`
     - `inapp` version updated to `8.6.0`
+- iOS
+    - Added support for Provisional Push.
 
 # 07-08-2024